[Bugfix] Fixed types for IDs in API calls

+++ ModCalls.cs
@@ -34,7 +34,7 @@ namespace OreExcavator
 							// Returns true or false if a specified Tile ID is whitelisted - or null if it's an invalid Tile ID
 							if (args.Length < 2)
 								throw new ArgumentException("Missing second argument: (int) Tile id");
-							if (args[1] is null || args[1] is not int tileID)
+							if (args[1] is null || args[1] is not ushort tileID)
 								throw new Exception($"Expected an argument of type (int) when getting a Tile ID, but got type ({args[1].GetType().Name}) instead.");
 							string name = GetFullNameById(tileID, ActionType.TileWhiteListed);
 							return ClientConfig.tileWhitelistToggled ? true : (name == null ? null : ClientConfig.tileWhitelist.Contains(name));
@@ -45,7 +45,7 @@ namespace OreExcavator
 							// Returns true or false if a specified Tile ID has been added successfully or not - or null if it exists already, or is an invalid Tile ID
 							if (args.Length < 2)
 								throw new ArgumentException("Missing second argument: (int) Tile id");
-							if (args[1] is null || args[1] is not int tileID)
+							if (args[1] is null || args[1] is not ushort tileID)
 								throw new Exception($"Expected an argument of type (int) when adding a Tile ID, but got type ({args[1].GetType().Name}) instead.");
 							string name = GetFullNameById(tileID, ActionType.TileWhiteListed);
 							if (name == null ? true : ClientConfig.tileWhitelist.Contains(name))
@@ -59,7 +59,7 @@ namespace OreExcavator
 							// Returns true or false if a specified Wall ID is whitelisted - or null if it's an invalid Wall ID
 							if (args.Length < 2)
 								throw new ArgumentException("Missing second argument: (int) Wall id");
-							if (args[1] is null || args[1] is not int wallID)
+							if (args[1] is null || args[1] is not ushort wallID)
 								throw new Exception($"Expected an argument of type (int) when getting a Wall ID, but got type ({args[1].GetType().Name}) instead.");
 							string name = GetFullNameById(wallID, ActionType.WallWhiteListed);
 							return ClientConfig.wallWhitelistToggled ? true : (name == null ? null : ClientConfig.wallWhitelist.Contains(name));
@@ -70,7 +70,7 @@ namespace OreExcavator
 							// Returns true or false if a specified Wall ID has been added successfully or not - or null if it exists already, or is an invalid Wall ID
 							if (args.Length < 2)
 								throw new ArgumentException("Missing second argument: (int) Wall id");
-							if (args[1] is null || args[1] is not int wallID)
+							if (args[1] is null || args[1] is not ushort wallID)
 								throw new Exception($"Expected an argument of type (int) when adding a Wall ID, but got type ({args[1].GetType().Name}) instead.");
 							string name = GetFullNameById(wallID, ActionType.WallWhiteListed);
 							if (name == null ? true : ClientConfig.wallWhitelist.Contains(name))
@@ -84,7 +84,7 @@ namespace OreExcavator
 							// Returns true or false if a specified Item ID is whitelisted - or null if it's an invalid Item ID
 							if (args.Length < 2)
 								throw new ArgumentException("Missing second argument: (int) Item id");
-							if (args[1] is null || args[1] is not int itemID)
+							if (args[1] is null || args[1] is not short itemID)
 								throw new Exception($"Expected an argument of type (int) when getting a Item ID, but got type ({args[1].GetType().Name}) instead.");
 							string name = GetFullNameById(itemID, ActionType.ItemWhiteListed);
 							return ClientConfig.itemWhitelistToggled ? true : (name == null ? null : ClientConfig.itemWhitelist.Contains(name));
@@ -95,7 +95,7 @@ namespace OreExcavator
 							// Returns true or false if a specified Item ID has been added successfully or not - or null if it exists already, or is an invalid Item ID
 							if (args.Length < 2)
 								throw new ArgumentException("Missing second argument: (int) Item id");
-							if (args[1] is null || args[1] is not int itemID)
+							if (args[1] is null || args[1] is not short itemID)
 								throw new Exception($"Expected an argument of type (int) when adding a Item ID, but got type ({args[1].GetType().Name}) instead.");
 							string name = GetFullNameById(itemID, ActionType.ItemWhiteListed);
 							if (name == null ? true : ClientConfig.itemWhitelist.Contains(name))
